Move hooks to separate package
