+++ shell.core/src/main/java/org/metaborg/spoofax/shell/client/ReplModule.java
@@ -5,6 +5,8 @@ import java.util.function.Consumer;
 import org.metaborg.spoofax.shell.commands.IReplCommand;
 import org.metaborg.spoofax.shell.commands.StyledText;
 import org.metaborg.spoofax.shell.core.CoreModule;
+import org.metaborg.spoofax.shell.hooks.OnEvalErrorHook;
+import org.metaborg.spoofax.shell.hooks.OnEvalSuccessHook;
 
 import com.google.inject.Singleton;
 import com.google.inject.TypeLiteral;
