+++ shell.core/src/main/java/org/metaborg/spoofax/shell/hooks/OnEvalSuccessHook.java
@@ -1,7 +1,8 @@
-package org.metaborg.spoofax.shell.client;
+package org.metaborg.spoofax.shell.hooks;
 
 import java.util.function.Consumer;
 
+import org.metaborg.spoofax.shell.client.IDisplay;
 import org.metaborg.spoofax.shell.commands.StyledText;
 
 import com.google.inject.Inject;
